Refactor stakepoolDataProvider
