+++ app/frontend/helpers/dataProviders/types.ts
@@ -1,3 +1,11 @@
+import {Stakepool} from '../../../frontend/types'
+
+export type StakepoolDataProvider = {
+  getPoolInfoByTicker: (ticker: string) => Stakepool
+  getPoolInfoByPoolHash: (poolHash: string) => Stakepool
+  hasTickerMapping: boolean
+}
+
 export type AssetDataProvider = {
   isADA: boolean
   isToken: boolean
